requirements.txt
matplotlib
numpy
pandas
pyqt5
pyvisa
